requirements.txt
fastapi
uvicorn
jinja2
azure-identity
azure-storage-blob
